modify unit tests failures.
